% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repint.R
\name{rep_int}
\alias{rep_int}
\title{ALTREP rep.int}
\usage{
rep_int(x, times)
}
\arguments{
\item{x}{the value}

\item{times}{how many times}
}
\description{
It only works with integers or values that can be converted to an integer
}
